requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
ollama>=0.1.0
python-dotenv>=1.0.0
click>=8.0.0
colorama>=0.4.6
prompt-toolkit>=3.0.0
pyyaml>=6.0
requests>=2.31.0
pyperclip>=1.8.0
